updated the 6-size.c file

+++ 0x00-hello_world/6-size.c
@@ -15,7 +15,7 @@ int main(void)
 	printf("Size of a char: %lu byte(s)\n", sizeof(ch));
 	printf("Size of a int: %lu byte(s)\n", sizeof(intType));
 	printf("Size of a long int: %lu byte(s)\n", sizeof(longType));
-	printf("Size of a long long int: %lu byte(s)\n", sizeof(doubleType));
+	printf("Size of a long long int: %lu byte(s)\n",(unsigned long)sizeof(doubleType));
 	printf("Size of a float: %lu byte(s)\n", sizeof(floatType));
 	return (0);
 }
